fixed bugs in migrate advice

+++ service/httpserver/controller/planController.go
@@ -85,12 +85,19 @@ func UserGetAdvice(con *gin.Context) {
 	if !valid {
 		return
 	}
+	user, infoSuccess := dao.UserDao.GetUserInfo(userID)
+	if !checkDaoSuccess(con, infoSuccess) {
+		return
+	}
 	advices, adviceSuccess := dao.MigrationAdviceDao.GetNewAdvice(userID)
 	if !checkDaoSuccess(con, adviceSuccess) {
 		return
 	}
-	log.Print("advices: ")
-	log.Println(*advices)
+	// choose advice
+	empty := make([]model.MigrationAdvice, 0)
+	if (*advices)[0].Status == args.AdviceStatusChoose && user.Status != args.UserForbiddenStatus {
+		advices = &empty
+	}
 	// return advices
 	con.JSON(http.StatusOK, gin.H{
 		"code": args.CodeOK,
@@ -115,18 +122,32 @@ func UserAbandonAdvice(con *gin.Context) {
 	if !valid {
 		return
 	}
-	deleteAdviceResult, deleteAdviceSuccess := dao.MigrationAdviceDao.DeleteAdvice(userID)
-	if !checkDaoSuccess(con, deleteAdviceSuccess) {
+	// delete advice with user status
+	user, infoSuccess := dao.UserDao.GetUserInfo(userID)
+	if !checkDaoSuccess(con, infoSuccess) {
 		return
 	}
-	// nothing has been delete
-	if deleteAdviceResult.DeletedCount == 0 {
+	if user.Status == args.UserForbiddenStatus {
 		con.JSON(http.StatusOK, gin.H{
-			"code": args.CodeDeleteNothing,
-			"msg":  "没有删除任何方案",
+			"code": args.CodeStatusForbidden,
+			"msg":  "用户正在迁移不能删除",
 			"data": gin.H{},
 		})
 		return
+	} else {
+		deleteAdviceResult, deleteAdviceSuccess := dao.MigrationAdviceDao.DeleteAdvice(userID)
+		if !checkDaoSuccess(con, deleteAdviceSuccess) {
+			return
+		}
+		// nothing has been delete
+		if deleteAdviceResult.DeletedCount == 0 {
+			con.JSON(http.StatusOK, gin.H{
+				"code": args.CodeDeleteNothing,
+				"msg":  "重复删除迁移建议",
+				"data": gin.H{},
+			})
+			return
+		}
 	}
 	con.JSON(http.StatusOK, gin.H{
 		"code": args.CodeOK,
@@ -238,10 +259,17 @@ func UserAcceptStoragePlan(con *gin.Context) {
 	if !checkDaoSuccess(con, adviceSuccess) {
 		return
 	}
+	// advice must be pending
 	nowAdvice := (*newAdvices)[0]
+	if nowAdvice.Status != args.AdviceStatusPending {
+		con.JSON(http.StatusOK, gin.H{
+			"code": args.CodeChangeNothing,
+			"msg":  "迁移建议非法,不能进行迁移",
+			"data": gin.H{},
+		})
+	}
 	// post to notice scheduler this plan
 	postPlanToSchedulerResponse, postPlanToSchedulerSuccess := scheduler.SetStoragePlanToScheduler(userID, code.AesDecrypt(user.Password, *args.EncryptKey), &nowAdvice.StoragePlanNew)
-	// save new plan
 	if !postPlanToSchedulerSuccess {
 		con.JSON(http.StatusOK, gin.H{
 			"code": args.CodeJsonError,
@@ -301,22 +329,13 @@ func UserAcceptStoragePlan(con *gin.Context) {
 		})
 		return
 	}
-	// delete advice
-	deleteAdviceResult, deleteAdviceSuccess := dao.MigrationAdviceDao.DeleteAdvice(userID)
-	if !checkDaoSuccess(con, deleteAdviceSuccess) {
+	// change advice status
+	_, setAdviceSuccess := dao.MigrationAdviceDao.SetAdviceStatus(userID, args.AdviceStatusChoose)
+	if !checkDaoSuccess(con, setAdviceSuccess) {
 		return
 	}
 	// recover user status : forbidden -> normal ?
 
-	// delete nothing
-	if deleteAdviceResult.DeletedCount == 0 {
-		con.JSON(http.StatusOK, gin.H{
-			"code": args.CodeDeleteNothing,
-			"msg":  "没有删除任何方案",
-			"data": gin.H{},
-		})
-		return
-	}
 	con.JSON(http.StatusOK, gin.H{
 		"code": args.CodeOK,
 		"msg":  "设置存储方案成功",
